refactor(PRODUCT): :sparkles: Updated usecase to inactivate
